COMPLITE films redux part
