+++ src/reducers/films/index.js
@@ -1,0 +1,4 @@
+export {default as filmsActions} from "./actions.js";
+export {default as filmsOperations} from "./operations.js";
+export {default as filmsSelectors} from "./selectors.js";
+export {default as filmsReducer} from "./reducers.js";
